Appointments now works
Hooray I actually did something right for a change wow call the
newspapers

Anyway appointments are done for users, it'll show all of the logged-in
user's appointments. I couldn't get the darn helpers.authorize to work
though, it got me trapped in a login loop upon trying to visit the page,
so I'm going to move onto the next thing and hopefully fix that tomorrow
if I've got time.